requirements.txt
pycryptodome==3.16.0
scrypt==0.8.20
tabulate==0.9.0
